+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed endpoint names for conflict event, food price, food security, population, poverty rate, and refugees tables
+- Include origin and asylum in data availability for refugees and returnees
 
 ## 0.9.6